feat: support dual proxy aggregation for notebooks and visualization
