+++ mcp/bin/mcp_proxy_bundle.cjs
