+++ mcp/package.json
