+++ mcp/tools/delete_cell.ts
