+++ mcp/tools/get_notebook_info.ts
