+++ mcp/tools/insert_cell.ts
