+++ mcp/tools/list_cells.ts
