+++ mcp/tools/read_cell.ts
