+++ mcp/tools/search_cells.ts
